requirements.txt
numpy>=1.20.0
scikit-learn==1.5.1
joblib>=1.1.0
pandas>=1.3.0
opencv-python>=4.5.0
scikit-image>=0.18.0
pillow>=8.0.0
scipy>=1.7.0 